Update capacity testings to new structure
