+++ package.json
@@ -27,17 +27,18 @@
     "test:performance:stress:browsing": "node __tests__/performance/stress/scripts/runStressFlow.js browsing",
     "test:performance:stress:orders": "node __tests__/performance/stress/scripts/runStressFlow.js orders",
     "test:performance:stress:payment": "node __tests__/performance/stress/scripts/runStressFlow.js payment",
-    "test:performance:capacity": "node __tests__/performance/stress/scripts/runAllCapacityFlows.js",
-    "test:performance:capacity:browsing": "node __tests__/performance/stress/scripts/runCapacityFlow.js browsing",
-    "test:performance:capacity:orders": "node __tests__/performance/stress/scripts/runCapacityFlow.js orders",
-    "test:performance:capacity:payment": "node __tests__/performance/stress/scripts/runCapacityFlow.js payment",
-    "test:performance:capacity:auth": "node __tests__/performance/stress/scripts/runCapacityFlow.js auth",
     "test:performance:spike": "node __tests__/performance/spike/scripts/runAllSpikeFlows.js",
     "test:performance:spike:auth-login": "node __tests__/performance/spike/scripts/runSpikeFlow.js auth.login",
     "test:performance:spike:auth-register": "node __tests__/performance/spike/scripts/runSpikeFlow.js auth.register",
     "test:performance:spike:browsing": "node __tests__/performance/spike/scripts/runSpikeFlow.js browsing",
     "test:performance:spike:orders": "node __tests__/performance/spike/scripts/runSpikeFlow.js orders",
-    "test:performance:spike:payment": "node __tests__/performance/spike/scripts/runSpikeFlow.js payment"
+    "test:performance:spike:payment": "node __tests__/performance/spike/scripts/runSpikeFlow.js payment",
+    "test:performance:capacity": "node __tests__/performance/capacity/scripts/runAllCapacityFlows.js",
+    "test:performance:capacity:browsing": "node __tests__/performance/capacity/scripts/runCapacityFlow.js browsing",
+    "test:performance:capacity:orders": "node __tests__/performance/capacity/scripts/runCapacityFlow.js orders",
+    "test:performance:capacity:payment": "node __tests__/performance/capacity/scripts/runCapacityFlow.js payment",
+    "test:performance:capacity:auth-login": "node __tests__/performance/capacity/scripts/runCapacityFlow.js auth.login",
+    "test:performance:capacity:auth-register": "node __tests__/performance/capacity/scripts/runCapacityFlow.js auth.register"
   },
   "keywords": [],
   "author": "RP",
